Don't use `grep -q` when operating on piped stdin
(By default, prefer `grep &>/dev/null`)

`grep -q` may exit as soon as it finds a match; this is a good optimization
for when the input is a file.  However, if the input is the output of
another program, then that other program will receive SIGPIPE, and further
writes will fail.  When this happens, it might (bsdtar does) print a
message about a "write error" to stderr.  Which is going to confuse and
alarm the user.

In one of the cases, this had already been mitigated by wrapping
bsdtar in "echo "$(bsdtar ...)", as Bash builtin echo doesn't complain
if it gets SIGPIPE.  However, that means we're storing the entire
output of bsdtar in memory, which is silly.

v2:
 - common.bash:checkRemovedPackageDB(): fix
 - db-functions:check_pkgfile: simplify; don't use grep at all
v3:
 - db-functions:check_pkgfile: use [[ = ]] instead of [[ == ]]