db-import-archlinux-src: Fixo
The first issue is a typo from last commit.  But I went ahead and tidied
the formatting around it.

The second is that create_tarballs() errored out if a repo doesn't contain
any arch=(any) packages.

The third is that it assumed that ABSLIBRE_ROOT and ABS_ROOT already exist.