test: common.bash: Be more careful about what we put in the package cache
The glob that copies makepkg outputs to the package cache was overly
permissive, and copies too much.  releasePackage calls __buildPackage
from "${TMP}/svn-packages-copy/${pkgbase}/trunk/".  This means that
when we update a package, after the updated version is built, both v1
and v2 exist in that trunk/ directory.  Then, this copies both the v1
and v2 .pkg.tar.xz files to the cache directory for v2.

That is: On subsequent runs (i.e. with a cache hit) of __buildPackage
for v2 of that package, it will look like makepkg magically built both
v1 and v2 of the package, instead of just v2.

Right now, that isn't really a problem because further down in
releasePackage, when we copy the makepkg outputs to $STAGING, that
glob is sufficiently strict to only copy the v2 outputs.  However, the
cache is poisoned, and that means that keeping it around longer
increases the odds that we do something that this poisoning breaks.

So, backport using libmakepkg's print_all_package_names to be
extremely careful about what we put in the cache.  This should make
the cache safe to use across test suite runs.

(this is a partial cherry pick of commit 76f95dd139094dc2985b6733f23fcb098f07c64e)

Allways build packages for all supported architectures