config: Avoid using && for flow-control
`set -e` doesn't trip on command lists containing &&, even if the command
list sets $? to non-zero.  However, when used as the last command in a
sourced file, that $? is used as the return value for the calling `source`
command, which then does trip `set -e`.

This affected `db-import-pkg`.

So, avoid the impropriety of spewing non-zero $? in the config file, since
doesn't know the context it's loaded in.