test: common.bash: checkPackageDB(): Require the DB file to exist
In 3a9623ba96dc8a26120be89354115dc120e71374 when I adjusted the function to
work correctly with `set -e`, I misread the interaction between parenthesis
and ||.  I had treated it like the more common
    ([ test ] && bsdtar) && fail    ->     if [ test ]; then ! bsdtar; fi
pattern, and incorrectly adjusted it as
    ([ test ] && bsdtar) || fail    ->     if [ test ]; then   bsdtar; fi
but I should have translated it as
    ([ test ] && bsdtar) || fail    ->     [ test ] && bsdtar

The result is that we weren't requiring the DB file to exist.

Note that Pierre got it right in Arch's version of the change,
d0b808b5a58023899c7e76958cdae732bbdfc08e.