rm -r db-check-nonfree-in-db tests-xihh/
`db-check-nonfree-in-db` (formerly `list_nonfree_in_db.py`) was used to
generate a list of packages to remove when importing a DB.  The import
scripts have not used it in a long time.  Delete it.

  (Following the above purpose, it shouldn't have been renamed to
  `db-check-*`; I mis-interpreted its purpose when I renamed it.)

For that matter, it probably should have been removed in

    commit f516cdb29d89027d267b3a09c98e518e7177251a
    Author: Nicolás Reynolds <fauno@kiwwwi.com.ar>
    Date:   Sat Oct 13 11:13:25 2012 -0300

        Deprecating old python scripts

By that commit, it was no longer used during import.  Further, it imports
from `filter.py`, which was removed in that commit.

While we're at it, remove `tests-xihh/` (which formerly lived in the main
`test/`).  The directory contains a test suite for the `filter.py` library
functions.  As we just established, `filter.py` no longer exists.