db-import-pkg: Don't do any work directly in $FTP_BASE/$repo or $WORKDIR/rsync
Currently, it
 - rsync's/creates (depending on $UPSTREAM) symlinks in
   "${FTP_BASE}/${repo}/os/${arch}" directly
 - runs repo-add on DBs in ${WORKDIR}/rsync/$(get_repo_dir "$repo" "$arch")
Instead, do both of these things in $WORKDIR/staging-rsync/$tag.  There's
no reason to re-use the initial /rsync/ dir (we don't re-use any files in
it--we rm them), and the sync_repo's --delete-excluded means that we
temporarily delete the .db files on the live repo.

Also, to do this efficiently, we need to create hardlinks between package
files in $FTP_BASE and $WORKDIR, so fiddle with tmp directories in the
config.

This also has the effect that imported repos no longer have '.db.old'
files, which is fine because they were never Parabola's old version of the
file, they were the verbatim .old file from where we're importing from.