db-import-pkg: Split make_repo_symlinks in 2
make_repo_symlinks() is doing 2 things:
 - resolving a package name to a file in /pool/
 - symlinking that file in /pool/ in to the staging dir
Split that in to 2 separate functions:
 - poolify(): take a list of packages, translate it in to a list of files
   in /pool/
 - make_repo_symlinks(): take that list of files, and symlink them in to
   the staging dir

While we're at it, this allows us to clean up the logic in poolify().  Part
of this is that it now takes the clean package list (from /new/), rather
than the glob-list that is .whitelist

This also means that it no-longer rewrites the .whitelist file.  Which is
exciting, because that was a confusing brain-f*ck.  The format of the file
magically changed half-way through the program... maybe! (depending on
the UPSTREAM).