db-import-*.conf: Split ARCHMIRROR in to ARCHMIRROR{,PATH}
Previously, db-import-pkg split ARCHMIRROR in to ARCHMIRROR_fullmodule and
ARCHMIRROR_path; such that

    ARCHMIRROR=${ARCHMIRROR_fullmodule}/${ARCHMIRROR_path}

This assumes that server gives the Arch mirror its own module; that isn't
always the case; for example in <rsync://ftp.acc.umu.se/mirror/archlinux/>
it lives in a subdirectory in the 'mirror' module.

So, we need a way to specify which part of the path leads to the root of
the mirror, and which part identifies the repo within it.  So, remove the
latter part from ARCHMIRROR, and create a separate ARCHMIRRORPATH variable
for it.

In db-import-pkg, this simply means that ARCHMIRROR_fullmodule becomes
ARCHMIRROR, and ARCHMIRROR_path becomes ARCHMIRRORPATH.