db-import-*.conf: Split off OUR{SRC,PKG}POOL from ARCH{SRC,PKG}POOL
Currently, db-import-pkg...
 - ... assumes that if a pkgpool is used on the upstream, we will be using
   the same pkgpool path; they are both configured with the single
   ARCHPKGPOOL variable.  Adjust it so that the local pkgpool path is
   configured with OURPKGPOOL.
 - ... decides whether the upstream uses an pkgpool and/or srcpool based on
   the upstream name.  Adjust it so that it decides this based on whether
   the ARCHPKGPOOL or ARCHSRCPOOL variables are set.

It's worth noting that this removes *almost* all of the logic that made
hard-coded decisions based on the upstream name; the only thing left is
that it still uses the upstream name to decide whether to let another
upstream's pkgpool override its arch=(any) packages.

"Ignore whitespace" might be useful when viewing this commit.