db-move, db-remove: Don't parse PKGBUILD files
Don't try to parse PKGBUILD files from XBS; all of the information we need
is already in the DBEXT files.  Several programs use [[ -f PKGBUILD ]] or
[[ -r PKGBUILD ]] on files from XBS; those checks can stay, just remove all
instances of actually trying to *parse* those files.

As an exception, don't modify parse_pkgbuilds.sh (which is called by
check_packages.py, which is called by cron-jobs/integrity-check).

Most of the attributes we need have been present in the DBEXT files for as
long as `repo-add` has been part of pacman:

    attribute |   git    |  git date  |  ver  |  ver date
    ----------+----------+------------+-------+-----------
    FILENAME  | aa1c0ba9 | 2006-11-20 | 3.0.0 | 2007-03-25
    NAME      | aa1c0ba9 | 2006-11-20 | 3.0.0 | 2007-03-25
    VERSION   | aa1c0ba9 | 2006-11-20 | 3.0.0 | 2007-03-25
    ARCH      | aa1c0ba9 | 2006-11-20 | 3.0.0 | 2007-03-25
    BASE      | 4b21504f | 2009-07-22 | 3.3.0 | 2009-08-02

However, `BASE` for split packages is slightly newer, and its presence
relies not only on the `repo-add` version, but also the `makepkg` version
used to build the package.

As of 2018-06-15, the oldest package in any of the Arch Linux repos is
community/gimp-refocus, built on 2013-07-22; I don't believe it is
necessary to handle packages from before that change was made (before
`BASE` was set).

"Ignore space change" might be useful when viewing this diff.