Begone with test/test.conf
Don't have the test architectures be dependent on the host architecture.
Count on qemu-user-static-binfmt making that OK.  If that isn't the case,
it's a bug in qemu-user-static-binfmt, and it should be fixed there, not
worked around here.

Do this in a way that reduces the diff size between Arch and us.